# Optimize .uniq() with just HashSet