--- a/crates/core/src/expr/array.rs
+++ b/crates/core/src/expr/array.rs
@@ -10,7 +10,7 @@ use anyhow::{Result, ensure};
 use reblessive::tree::Stk;
 use revision::revisioned;
 use serde::{Deserialize, Serialize};
-use std::collections::HashSet;
+use std::collections::HashMap;
 use std::fmt::{self, Display, Formatter, Write};
 use std::ops;
 use std::ops::Deref;
@@ -437,19 +437,16 @@ pub(crate) trait Uniq<T> {
 }
 
 impl Uniq<Array> for Array {
-	fn uniq(mut self) -> Array {
+	fn uniq(self) -> Array {
 		#[expect(clippy::mutable_key_type)]
-		let mut set: HashSet<&Value> = HashSet::new();
-		let mut to_remove: Vec<usize> = Vec::new();
-		for (i, item) in self.iter().enumerate() {
-			if !set.insert(item) {
-				to_remove.push(i);
+		let mut map: HashMap<&Value, ()> = HashMap::with_capacity(self.len());
+		let mut to_return = Array::with_capacity(self.len());
+		for i in self.iter() {
+			if map.insert(i, ()).is_some() {
+				to_return.push(i.clone());
 			}
 		}
-		for i in to_remove.iter().rev() {
-			self.remove(*i);
-		}
-		self
+		to_return
 	}
 }