--- a/crates/core/src/expr/array.rs
+++ b/crates/core/src/expr/array.rs
@@ -10,7 +10,7 @@ use anyhow::{Result, ensure};
 use reblessive::tree::Stk;
 use revision::revisioned;
 use serde::{Deserialize, Serialize};
-use std::collections::HashMap;
+use std::collections::HashSet;
 use std::fmt::{self, Display, Formatter, Write};
 use std::ops;
 use std::ops::Deref;
@@ -439,10 +439,10 @@ pub(crate) trait Uniq<T> {
 impl Uniq<Array> for Array {
 	fn uniq(self) -> Array {
 		#[expect(clippy::mutable_key_type)]
-		let mut map: HashMap<&Value, ()> = HashMap::with_capacity(self.len());
+		let mut set = HashSet::with_capacity(self.len());
 		let mut to_return = Array::with_capacity(self.len());
 		for i in self.iter() {
-			if map.insert(i, ()).is_some() {
+			if set.insert(i) {
 				to_return.push(i.clone());
 			}
 		}