Thank you for submitting this pull request. We really appreciate you spending the time to work on SurrealDB. 🚀 🎉 

## What is the motivation?

The current implementation of .uniq() seemed like it could be simplified and optimized with just a HashSet instead of also tracking indexes and then using .remove().

## What does this change do?

It checks to see if an item is already present, and if it is not then it pushes it into an array. Then the array is returned with only the unique items.

## What is your testing strategy?

Added a Criterion benchmark (included BTreeSet for comparison to show that it is worse). On my machine `cargo bench --bench other_uniq_implementations` shows this sort of output:

```
current_uniq            time:   [11.999 ms 12.039 ms 12.084 ms]
uniq_hashset            time:   [8.5845 ms 8.6044 ms 8.6261 ms]
uniq_btreeset           time:   [36.850 ms 36.922 ms 36.993 ms]
```

## Is this related to any issues?

- [X] No related issues

## Does this change need documentation?

- [X] No documentation needed

## Does this change make any alterations to environment variables or CLI commands?

- [X] No changes made to env vars

## Have you read the Contributing Guidelines?

- [X] I have read the [Contributing Guidelines](https://github.com/surrealdb/surrealdb/blob/main/CONTRIBUTING.md)